Move src/css_util.ts to src/lib/
